Create the alternate localisation algorithm
Added the basis for the second method that we will be testing
for localising on multiple floors.

Each floor will run MCL in parallel, then send information about
the largest cluster to the decider node.

The decider node looks at the clusters coming from each floor, and
at the moment naively chooses the floor with the largest cluster.